LayoutTests/http/tests/security/no-javascript-refresh-spaces-expected.txt

CONSOLE MESSAGE: line 1: Refused to refresh http://127.0.0.1:8000/security/no-javascript-refresh-spaces.php to a javascript: URL
PASS: This is the content that appears in place of a refresh.
